+++ History.md
@@ -19,6 +19,7 @@
 - When exporting the proposed procedure list into ODS, there is now an option to also include the comments.
 - The admin interface for adding/removing supporters of a motion/amendment now has a function to copy the full list of supporters to the clipboard in a format suitable to paste it into the full-text field later on (to easily transfer the supporter list from one motion to a new one).
 - The diff view in amendments now usually show the whole affected line, instead of cutting off the line after the last changed word, giving some more context to the change.
+- Add a option to enforce a confirmation checkbox on registration.
 - Bugfix: When a motion type was created from the scratch and motion sections positioned right were added, the layout didn't switch to two-column mode.
 - Bugfix: When an optional image was not uploaded, the LaTeX-based PDF export did not work.
 - Bugfix: It was not possible to delete admin comments in the proposed procedure.